pyproject.toml
[tool.poetry]
name = "jacie"
version = "0.1.0"
description = ""
authors = ["elroy-galbraith <elroy@yoii.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain-community = "^0.3.18"
faiss-cpu = "^1.10.0"
langchain-google-vertexai = "^2.0.13"
streamlit = "^1.42.2"
pypdf = "^5.3.0"
langchain-text-splitters = "^0.3.6"
pdfplumber = "^0.11.5"
fitz = "^0.0.1.dev2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
